+++ CHANGELOG.md
@@ -8,7 +8,14 @@ and this project adheres to
 
 ## [Unreleased]
 
-## 5.2.0 - 2020-09-21
+## 5.3.1 - 2020-09-21
+
+### Fixed
+
+- Removed `ingestResourceManager` which caused previously-configured
+  integrations to stop ingesting resource manager steps
+
+## 5.3.0 - 2020-09-21
 
 - Added `azure_service_bus_namespace` entities
 - Added `azure_resource_group|has|azure_service_bus_namespace` relationships